+++ tests/testthat/setup-files.R
@@ -1,29 +1,42 @@
+library(here)
 library(logger)
 
 log_threshold(ERROR)
 
 TXTPATH <- tempfile(fileext = ".txt")
-PNGPATH <- tempfile(fileext = ".png")
+CSSPATH <- tempfile(fileext = ".css")
 HTMLPATH <- "hello.html"
 RMD_TEMPLATE <- "vignette.Rmd"
+
+# This file comes from https://bit.ly/2P4LUO8 (cat poster on WikiPedia).
+#
+JPGPATH <- system.file("cats.jpg", package = "emayili", mustWork = TRUE)
+# If have not installed yet.
+if (!file.exists(JPGPATH)) JPGPATH <- here("inst", "cats.jpg")
+
+TXTCONTENT <- "Some random text."
+
+PLAIN_MARKDOWN <- "## Section\n[link](https://www.google.com)"
+PLAIN_MARKDOWN_INTERPOLATE <- "Hi {{name}}!"
+
+COLOUR_GLAUCOUS = "#6082b6"
 
 # Start with a blank slate.
 #
 source("teardown-files.R", local = TRUE)
 
-writeLines("Some random text.", TXTPATH)
+# The , sep = "" prevents it from writing an "\n" at the end of the line.
+writeLines(TXTCONTENT, TXTPATH, sep = "")
+TXTCONTENT_ENCODED <- emayili:::mime_base64encode(charToRaw(TXTCONTENT))
 
-png(PNGPATH, width=600, height=350)
-hist(mtcars$disp)
-dev.off()
+writeLines(
+  paste0("body {color: ", COLOUR_GLAUCOUS," !important;}"),
+  CSSPATH
+)
 
 HTMLCONTENT <- "<p>Hello there, stranger!</p>"
 #
 writeLines(HTMLCONTENT, HTMLPATH)
-
-# This file comes from https://bit.ly/2P4LUO8 (cat poster on WikiPedia).
-#
-JPGPATH <- here::here("inst", "cats.jpg")
 
 # SERVER -----------------------------------------------------------------------
 
@@ -34,6 +47,8 @@ JPGPATH <- here::here("inst", "cats.jpg")
 #
 SMTP_SERVER   = "mail.smtpbucket.com"
 SMTP_PORT     = 8025
+SMTP_USERNAME <- Sys.getenv("SMTP_USERNAME")
+SMTP_PASSWORD <- Sys.getenv("SMTP_PASSWORD")
 
 smtp <- server(
   host = SMTP_SERVER,
@@ -68,8 +83,3 @@ rmarkdown::draft(
   package = "rmarkdown",
   edit = FALSE
 )
-
-# ------------------------------------------------------------------------------
-
-SMTP_USERNAME <- Sys.getenv("SMTP_USERNAME")
-SMTP_PASSWORD <- Sys.getenv("SMTP_PASSWORD")