+++ tests/testthat/test-attachment.R
@@ -1,22 +1,28 @@
 test_that("attachment: set", {
   msg <- envelope() %>% attachment(TXTPATH)
 
-  expect_equal(msg$parts[[1]]$content, base64encode(TXTPATH, 76L, "\r\n"))
+  expect_match(as.character(msg), TXTCONTENT_ENCODED)
 })
 
 test_that("attachment: specify CID", {
   cid <- "histogram"
 
-  msg <- envelope() %>% attachment(PNGPATH, cid = cid)
+  msg <- envelope() %>% attachment(JPGPATH, cid = cid)
 
-  expect_equal(msg$parts[[1]]$content, base64encode(PNGPATH, 76L, "\r\n"))
-  expect_equal(msg$parts[[1]]$cid, cid)
+  expect_true(
+    grepl(
+      mime_base64encode(JPGPATH),
+      as.character(msg),
+      fixed = TRUE
+    )
+  )
+  expect_equal(msg$parts$cid, cid)
 })
 
 test_that("attachment: number of files", {
   msg <- envelope()
 
   expect_error(msg %>% attachment())
-  expect_error(msg %>% attachment(c(TXTPATH, PNGPATH)))
+  expect_error(msg %>% attachment(c(TXTPATH, JPGPATH)))
 })